3rd attempt

+++ set_1.py
@@ -1,5 +1,5 @@
 def savings(gross_pay, tax_rate, expenses):
-    return ((gross_pay - int(gross_pay * tax_rate))- expenses)
+    return int((gross_pay - int(gross_pay * tax_rate))- expenses)
 def material_waste(total_material, material_units, num_jobs, job_consumption):
     return str(total_material - (num_jobs * job_consumption)) + str(material_units)
 def interest(principal, rate, periods):
